Dial back the frequency of releases cache updates
We don't really need to update every 1 minute. Every 5 minutes is sufficient.
I've noticed some latency issues with getting responses from the API, which I'm
assuming is because it's in the middle of updating the cache when I send my
request. Hopefully, dialing it back to once every 5 minutes will help. The
downside is that when I deploy a new version of Alda (or a change to an existing
version), the change won't be represented in the API for up to 5 minutes. But, I
think that's acceptable.